webui: start reorganizing the component structure
