+++ webui/src/bug/LabelChange.tsx
@@ -1,9 +1,9 @@
 import { makeStyles } from '@material-ui/core/styles';
 import React from 'react';
 
-import Author from '../Author';
-import Date from '../Date';
-import Label from '../Label';
+import Author from '../components/Author';
+import Date from '../components/Date';
+import Label from '../components/Label';
 
 import { LabelChangeFragment } from './LabelChangeFragment.generated';
