+++ webui/src/bug/MessageCommentFragment.graphql
@@ -1,4 +1,4 @@
-#import "../Author.graphql"
+#import "../components/fragments.graphql"
 
 fragment AddComment on AddCommentTimelineItem {
   createdAt
