+++ webui/src/bug/MessageCreateFragment.graphql
@@ -1,4 +1,4 @@
-#import "../Author.graphql"
+#import "../components/fragments.graphql"
 
 fragment Create on CreateTimelineItem {
   createdAt
