+++ webui/src/bug/SetStatusFragment.graphql
@@ -1,4 +1,4 @@
-#import "../Author.graphql"
+#import "../components/fragments.graphql"
 
 fragment SetStatus on SetStatusTimelineItem {
   date
