+++ webui/src/bug/SetTitleFragment.graphql
@@ -1,4 +1,4 @@
-#import "../Author.graphql"
+#import "../components/fragments.graphql"
 
 fragment SetTitle on SetTitleTimelineItem {
   date
